refactor: split logic to utils files
